Revert "Merge branch 'develop' into origin/UI/Game/HUD-IngameMenu"
This reverts commit d50bfe1a5773080a07b5354fecdda0a8faf1c3c6, reversing
changes made to da5db03615d7a1c657f56bd7adf62527ebf25b85.

+++ Source/AxoLog.h
@@ -67,7 +67,7 @@ private:
 	{
 		LogSeverity severity;
 		std::string file;
-		uint16_t line;
+		unsigned int line;
 		std::string message;
 
 		std::string ToDetailedString(bool addBreak = true) const;
