+++ Source/DataModels/Windows/EditorWindows/WindowStateMachineEditor.cpp
@@ -83,9 +83,9 @@ void WindowStateMachineEditor::DrawWindowContents()
 	DrawGridStateMachine(canvasP0, scrolling, io.MouseDelta, drawList);
 	
 	ImGui::SetCursorScreenPos(ImVec2(canvasP0.x + 15, canvasP0.y + 50));
-	int pos = static_cast<int>(mousePosInCanvas.x);
+	int pos = mousePosInCanvas.x;
 	std::string positionString = std::to_string(pos) + ",";
-	pos = static_cast<int>(mousePosInCanvas.y);
+	pos = mousePosInCanvas.y;
 	positionString += std::to_string(pos);
 	ImGui::Text(positionString.c_str());
 
@@ -111,10 +111,7 @@ void WindowStateMachineEditor::DrawWindowContents()
 		{
 			std::pair<int, int> posSelectedState = stateAsShared->GetState(stateIdSelected)->auxiliarPos;
 			DrawActualTransitionCreation(
-				ImVec2(static_cast<float>(posSelectedState.first), static_cast<float>(posSelectedState.second)),
-				origin,
-				mousePosInCanvas,
-				drawList);
+				ImVec2(posSelectedState.first, posSelectedState.second), origin, mousePosInCanvas, drawList);
 		}
 
 		DrawTransitions(stateAsShared, origin, drawList);
@@ -224,7 +221,7 @@ void WindowStateMachineEditor::DrawParameters(std::shared_ptr<ResourceStateMachi
 
 void WindowStateMachineEditor::DrawStateEditor(std::shared_ptr<ResourceStateMachine>& stateAsShared)
 {
-	if (stateIdSelected < static_cast<int>(stateAsShared->GetNumStates()) && stateIdSelected > 0)
+	if (stateIdSelected < stateAsShared->GetNumStates() && stateIdSelected > 0)
 	{
 		State* state = stateAsShared->GetState(stateIdSelected);
 		ImGui::Text("State");
@@ -243,7 +240,7 @@ void WindowStateMachineEditor::DrawStateEditor(std::shared_ptr<ResourceStateMach
 		else
 		{
 			ImGui::Checkbox("Loop", &state->loop);
-			ImGui::DragFloat("Speed", &state->speed, 0.25f, 0.1f, 100.f, "%.1f");
+			ImGui::DragFloat("Speed", &state->speed, 0.25f, 0.1, 100, "%.1f");
 
 			if (ImGui::Button("x"))
 			{
@@ -423,7 +420,7 @@ void WindowStateMachineEditor::DrawTransitions(std::shared_ptr<ResourceStateMach
 		Transition transition = it.second;
 		State* stateOrigin = stateAsShared->GetState(transition.originState);
 		State* stateDestination = stateAsShared->GetState(transition.destinationState);
-		ImGui::PushID(static_cast<int>(it.first));
+		ImGui::PushID(it.first);
 		ImVec2 posStateOriginCenter = ImVec2(origin.x + stateOrigin->auxiliarPos.first + sizeState.x / 2,
 											 origin.y + stateOrigin->auxiliarPos.second + sizeState.y / 2);
 		ImVec2 postStateDestinationCenter = ImVec2(origin.x + stateDestination->auxiliarPos.first + sizeState.x / 2,
@@ -498,14 +495,13 @@ void WindowStateMachineEditor::DrawStates(std::shared_ptr<ResourceStateMachine>&
 										  const ImVec2& mouseDelta,
 										  ImDrawList* drawList)
 {
-	for (unsigned int i = 0; i < stateAsShared->GetNumStates(); i++)
+	for (int i = 0; i < stateAsShared->GetNumStates(); i++)
 	{
 		State* state = stateAsShared->GetState(i);
 		if (state != nullptr)
 		{
 			ImGui::PushID(i);
-			ImVec2 posState =
-				ImVec2(static_cast<float>(state->auxiliarPos.first), static_cast<float>(state->auxiliarPos.second));
+			ImVec2 posState = ImVec2(state->auxiliarPos.first, state->auxiliarPos.second);
 			ImVec2 minRect = ImVec2(origin.x + posState.x, origin.y + posState.y);
 			ImVec2 maxRect = ImVec2(origin.x + posState.x + sizeState.x, origin.y + posState.y + sizeState.y);
 			ImGui::SetCursorScreenPos(minRect);
@@ -536,8 +532,7 @@ void WindowStateMachineEditor::DrawStates(std::shared_ptr<ResourceStateMachine>&
 				}
 				if (ImGui::IsMouseDragging(ImGuiMouseButton_Left))
 				{
-					state->auxiliarPos = std::make_pair<int, int>(static_cast<int>(posState.x + mouseDelta.x),
-																  static_cast<int>(posState.y + mouseDelta.y));
+					state->auxiliarPos = std::make_pair<int, int>(posState.x + mouseDelta.x, posState.y + mouseDelta.y);
 					// stateAsShared->EditState(i, state);
 				}
 				if (ImGui::IsMouseReleased(ImGuiMouseButton_Right))
@@ -615,7 +610,7 @@ void WindowStateMachineEditor::DrawRightClickPopUp(std::shared_ptr<ResourceState
 		openContextMenu = false;
 		if (stateAsShared)
 		{
-			if (stateIdSelected >= 0 && stateIdSelected < static_cast<int>(stateAsShared->GetNumStates()))
+			if (stateIdSelected >= 0 && stateIdSelected < stateAsShared->GetNumStates())
 			{
 				if (ImGui::MenuItem("Add Transition"))
 				{
@@ -641,8 +636,7 @@ void WindowStateMachineEditor::DrawRightClickPopUp(std::shared_ptr<ResourceState
 			{
 				if (ImGui::MenuItem("Create State"))
 				{
-					stateAsShared->AddNewState(static_cast<int>(mousePosInCanvas.x),
-											   static_cast<int>(mousePosInCanvas.y));
+					stateAsShared->AddNewState(mousePosInCanvas.x, mousePosInCanvas.y);
 				}
 			}
 		}
