+++ Source/FileSystem/Importers/AnimationImporter.cpp
@@ -42,7 +42,7 @@ void AnimationImporter::Load(const char* fileBuffer, std::shared_ptr<ResourceAni
 	std::unordered_map<std::string, ResourceAnimation::Channel*> channels;
 	channels.reserve(header[0]);
 
-	for (unsigned int i = 0; i < header[0]; ++i)
+	for (int i = 0; i < header[0]; ++i)
 	{
 		ResourceAnimation::Channel* channel = new ResourceAnimation::Channel;
 		std::string channelName;
@@ -82,16 +82,18 @@ void AnimationImporter::Load(const char* fileBuffer, std::shared_ptr<ResourceAni
 
 void AnimationImporter::Save(const std::shared_ptr<ResourceAnimation>& resource, char*& fileBuffer, unsigned int& size)
 {
-	unsigned int header[1] = { resource->GetNumChannels() };
+	unsigned int header[1] = {
+		resource->GetNumChannels()
+	};
 
 	size = (sizeof(unsigned int) * 3) * resource->GetNumChannels() + sizeof(header) + sizeof(double);
 
 	for (auto it : resource->GetChannels())
 	{
 		ResourceAnimation::Channel* channel = it.second;
-		size += sizeof(float3) * static_cast<unsigned int>(channel->positions.size());
-		size += sizeof(Quat) * static_cast<unsigned int>(channel->rotations.size());
-		size += sizeof(char) * static_cast<unsigned int>(it.first.length());
+		size += sizeof(float3) * channel->positions.size();
+		size += sizeof(Quat) * channel->rotations.size();
+		size += sizeof(char) * it.first.length();
 	}
 
 	char* cursor = new char[size]{};
@@ -115,9 +117,9 @@ void AnimationImporter::Save(const std::shared_ptr<ResourceAnimation>& resource,
 		ResourceAnimation::Channel* channel = it.second;
 
 		unsigned int nodeHeader[3] = {
-			static_cast<unsigned int>(it.first.length()),
-			static_cast<unsigned int>(channel->positions.size()),
-			static_cast<unsigned int>(channel->rotations.size()),
+			it.first.length(),
+			channel->positions.size(),
+			channel->rotations.size(),
 		};
 
 		bytes = sizeof(nodeHeader);
@@ -125,14 +127,14 @@ void AnimationImporter::Save(const std::shared_ptr<ResourceAnimation>& resource,
 
 		cursor += bytes;
 
-		bytes = sizeof(char) * static_cast<unsigned int>(it.first.length());
+		bytes = sizeof(char) * it.first.length();
 		memcpy(cursor, it.first.c_str(), bytes);
 
 		cursor += bytes;
 
 		if (!channel->positions.empty())
 		{
-			bytes = sizeof(float3) * static_cast<unsigned int>(channel->positions.size());
+			bytes = sizeof(float3) * channel->positions.size();
 			memcpy(cursor, &(channel->positions[0]), bytes);
 
 			cursor += bytes;
@@ -140,7 +142,7 @@ void AnimationImporter::Save(const std::shared_ptr<ResourceAnimation>& resource,
 
 		if (!channel->rotations.empty())
 		{
-			bytes = sizeof(Quat) * static_cast<unsigned int>(channel->rotations.size());
+			bytes = sizeof(Quat) * channel->rotations.size();
 			memcpy(cursor, &(channel->rotations[0]), bytes);
 
 			cursor += bytes;
